Rename cavl.h -> _canard_cavl.h, fix https://github.com/OpenCyphal/libcanard/issues/196
